+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ---
 
 
+### v1.0.3 (30/6/2023)
+
+---
+
+- [x] Bug fixes for media meta components
+
+- [x] Add support for documents upload/document folders
+
+
+
 ### v1.0.2 (30/6/2023)
 
 ---